backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
google-generativeai
python-multipart
ultralytics
opencv-python-headless
numpy
